requirements.txt
arxiv
google-genai
pydantic 